backend/requirements.txt
# Core Dependencies
motor>=3.4.0
beanie==1.25.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.6
authlib==1.2.1
httpx==0.25.2
python-dotenv==1.0.0
certifi>=2023.11.17
requests
email-validator>=2.0.0

# RAG & Document Processing
faiss-cpu==1.7.4
numpy==1.26.2
pypdf==3.17.1
pdfplumber==0.10.3
python-docx==1.1.0
python-pptx==1.0.2
openpyxl==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
pillow==10.1.0
# Removed tiktoken as it is unused

# Removed unused/heavy:
# langchain, langchain-community, google-generativeai, ollama, pandas, scikit-learn